+++ frontend/src/components/appBar/AppBar.tsx
@@ -1,18 +1,12 @@
+import React from "react";
 import Link from "next/link";
-import IconButton from "@mui/material/IconButton";
-import ExitToAppIcon from "@mui/icons-material/ExitToApp";
-import { useMutation, useQuery } from "@apollo/client";
-import { UserContextTypes } from "@/types/UserTypes";
+import { useRouter } from "next/router";
 import {
   AppBar,
-  Avatar,
   Box,
   Button,
   Container,
-  Menu,
-  MenuItem,
   Toolbar,
-  Tooltip,
   Typography,
 } from "@mui/material";
 import CookieIcon from "@mui/icons-material/Cookie";
@@ -20,12 +14,8 @@ import EditNoteIcon from "@mui/icons-material/EditNote";
 import SearchIcon from "@mui/icons-material/Search";
 import ContactSupportIcon from "@mui/icons-material/ContactSupport";
 import AccountCircleIcon from "@mui/icons-material/AccountCircle";
-import LoginIcon from "@mui/icons-material/Login";
-import MenuIcon from "@mui/icons-material/Menu";
-import React, { useEffect, useState } from "react";
-import { useRouter } from "next/router";
-import { PATH_IMAGE } from "@/api/configApi";
-import { mutationSignOut, queryMeContext } from "../graphql/Users";
+import BurgerMenu from "./BurgerMenu";
+import UserMenu from "./UserMenu";
 
 const buttonStyles = {
   color: "white",
@@ -36,62 +26,12 @@ const buttonStyles = {
 
 export default function Header(): React.ReactNode {
   const router = useRouter();
-  // User connected ?
-  const { data, error } = useQuery<{ item: UserContextTypes }>(queryMeContext);
-  const [userContext, setUserContext] = useState<UserContextTypes>(null);
-  const [userConnected, setUserConnected] = useState<Boolean>(false);
 
-  useEffect(() => {
-    if (error) {
-      setUserContext(null);
-      setUserConnected(false);
-    }
-    if (data?.item) {
-      setUserContext(data.item);
-      setUserConnected(true);
-    }
-  }, [data, error]);
-
-  // Signout
-  const [doSignout] = useMutation(mutationSignOut, {
-    onCompleted: () => {
-      setUserContext(null);
-      setAnchorElUser(null);
-      setUserConnected(false);
-      router.replace(`/signin`);
-    },
-    refetchQueries: [{ query: queryMeContext }],
-  });
-  async function logout() {
-    doSignout();
-  }
-
-  // Open / Close menu
-  const [anchorElNav, setAnchorElNav] = React.useState<null | HTMLElement>(
-    null,
-  );
-  const [anchorElUser, setAnchorElUser] = React.useState<null | HTMLElement>(
-    null,
-  );
-
-  const handleOpenNavMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorElNav(event.currentTarget);
-  };
-  const handleOpenUserMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorElUser(event.currentTarget);
-  };
-
-  const handleCloseNavMenu = () => {
-    setAnchorElNav(null);
-  };
-
-  const handleCloseUserMenu = () => {
-    setAnchorElUser(null);
-  };
   return (
     <AppBar position="static" sx={{ backgroundColor: "#343a40" }}>
       <Container maxWidth="xl">
         <Toolbar disableGutters className="header">
+          {/* Desktop Title */}
           <Box
             className="header_title"
             sx={{
@@ -105,71 +45,9 @@ export default function Header(): React.ReactNode {
               THE GOOD COOKIE
             </Link>
           </Box>
-          <Box sx={{ flexGrow: 1, display: { xs: "flex", md: "none" } }}>
-            <IconButton
-              size="large"
-              aria-label="account of current user"
-              aria-controls="menu-appbar"
-              aria-haspopup="true"
-              onClick={handleOpenNavMenu}
-              sx={{
-                color: "#e89116",
-              }}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Menu
-              id="menu-appbar"
-              anchorEl={anchorElNav}
-              anchorOrigin={{
-                vertical: "bottom",
-                horizontal: "left",
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: "top",
-                horizontal: "left",
-              }}
-              open={Boolean(anchorElNav)}
-              onClose={handleCloseNavMenu}
-              sx={{
-                display: { xs: "block", md: "none" },
-              }}
-            >
-              <MenuItem
-                onClick={() => {
-                  handleCloseNavMenu();
-                  router.replace(`/annonces/new`);
-                }}
-              >
-                <Typography textAlign="center"> Déposer une annonce</Typography>
-              </MenuItem>
-              <MenuItem
-                onClick={() => {
-                  handleCloseNavMenu();
-                  router.replace(`/recherche`);
-                }}
-              >
-                <Typography textAlign="center"> Recherche</Typography>
-              </MenuItem>
-              <MenuItem
-                onClick={() => {
-                  handleCloseNavMenu();
-                  router.replace(`/compte`);
-                }}
-              >
-                <Typography textAlign="center"> Compte</Typography>
-              </MenuItem>
-              <MenuItem
-                onClick={() => {
-                  handleCloseNavMenu();
-                  router.replace(`/contact`);
-                }}
-              >
-                <Typography textAlign="center"> Contact</Typography>
-              </MenuItem>
-            </Menu>
-          </Box>
+          {/* Mobile Menu */}
+          <BurgerMenu />
+          {/* Mobile Title */}
           <CookieIcon
             sx={{
               display: { xs: "flex", md: "none" },
@@ -180,8 +58,7 @@ export default function Header(): React.ReactNode {
           <Typography
             variant="h4"
             noWrap
-            component="a"
-            href="/"
+            component="div"
             sx={{
               mr: 2,
               display: { xs: "flex", md: "none" },
@@ -192,14 +69,15 @@ export default function Header(): React.ReactNode {
               textDecoration: "none",
             }}
           >
-            TGC
+            <Link href="/" passHref className="header_title_name">
+              TGC
+            </Link>
           </Typography>
           <Box sx={{ flexGrow: 1, display: { xs: "none", md: "flex" } }}>
             <Button
               className="header_link_button"
               startIcon={<EditNoteIcon fontSize="large" />}
               onClick={() => {
-                handleCloseNavMenu();
                 router.replace(`/annonces/new`);
               }}
               sx={buttonStyles}
@@ -210,7 +88,6 @@ export default function Header(): React.ReactNode {
               className="header_link_button"
               startIcon={<SearchIcon />}
               onClick={() => {
-                handleCloseNavMenu();
                 router.replace(`/recherche`);
               }}
               sx={buttonStyles}
@@ -221,7 +98,6 @@ export default function Header(): React.ReactNode {
               className="header_link_button"
               startIcon={<AccountCircleIcon />}
               onClick={() => {
-                handleCloseNavMenu();
                 router.replace(`/compte`);
               }}
               sx={buttonStyles}
@@ -232,7 +108,6 @@ export default function Header(): React.ReactNode {
               className="header_link_button"
               startIcon={<ContactSupportIcon />}
               onClick={() => {
-                handleCloseNavMenu();
                 router.replace(`/contact`);
               }}
               sx={buttonStyles}
@@ -240,66 +115,7 @@ export default function Header(): React.ReactNode {
               contact
             </Button>
           </Box>
-
-          <Box sx={{ flexGrow: 0 }}>
-            <Tooltip title="Ouvrir le menu du profil">
-              <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
-                <Avatar
-                  alt="User avatar"
-                  src={
-                    userConnected && userContext.picture.path
-                      ? `${userContext.picture.path}`
-                      : `${PATH_IMAGE}/default/avatar.webp`
-                  }
-                />
-              </IconButton>
-            </Tooltip>
-
-            <Menu
-              sx={{ mt: "45px" }}
-              id="menu-appbar"
-              anchorEl={anchorElUser}
-              anchorOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              open={Boolean(anchorElUser)}
-              onClose={handleCloseUserMenu}
-            >
-              {userConnected ? (
-                <Box>
-                  <MenuItem
-                    onClick={() => {
-                      handleCloseUserMenu();
-                      router.replace(`/compte`);
-                    }}
-                  >
-                    <AccountCircleIcon />
-                    <Typography textAlign="center">Mon compte</Typography>
-                  </MenuItem>
-                  <MenuItem onClick={logout}>
-                    <ExitToAppIcon />
-                    <Typography textAlign="center">Se déconnecter</Typography>
-                  </MenuItem>
-                </Box>
-              ) : (
-                <MenuItem
-                  onClick={() => {
-                    handleCloseUserMenu();
-                    router.replace(`/signin`);
-                  }}
-                >
-                  <LoginIcon />
-                  <Typography textAlign="center">Connexion</Typography>
-                </MenuItem>
-              )}
-            </Menu>
-          </Box>
+          <UserMenu />
         </Toolbar>
       </Container>
     </AppBar>
